WIP. Major object creation and configuration refactoring
