+++ libraries/Authentication/microsoft-agents-authentication-msal/microsoft/agents/authentication/msal/__init__.py
@@ -1,9 +1,5 @@
-from .auth_types import AuthTypes
-from .msal_auth_configuration import MsalAuthConfiguration
 from .msal_auth import MsalAuth
 
 __all__ = [
-    "AuthTypes",
-    "MsalAuthConfiguration",
     "MsalAuth",
 ]
