--- libraries/Authentication/microsoft-agents-authentication-msal/microsoft/agents/authentication/msal/msal_auth_configuration.py
@@ -1,42 +1,0 @@
-from typing import Optional
-
-from microsoft.agents.authorization import AgentAuthConfiguration
-
-from .auth_types import AuthTypes
-
-
-class MsalAuthConfiguration(AgentAuthConfiguration):
-    """
-    Configuration for MSAL authentication.
-    """
-
-    AUTH_TYPE: AuthTypes
-    CLIENT_SECRET: Optional[str]
-    CERT_PEM_FILE: Optional[str]
-    CERT_KEY_FILE: Optional[str]
-    CONNECTION_NAME: Optional[str]
-    SCOPES: Optional[list[str]]
-    AUTHORITY: Optional[str]
-
-    def __init__(
-        self,
-        auth_type: AuthTypes = None,
-        client_id: str = None,
-        tenant_id: Optional[str] = None,
-        client_secret: Optional[str] = None,
-        cert_pem_file: Optional[str] = None,
-        cert_key_file: Optional[str] = None,
-        connection_name: Optional[str] = None,
-        authority: Optional[str] = None,
-        scopes: Optional[list[str]] = None,
-        **kwargs: Optional[dict[str, str]],
-    ):
-        self.AUTH_TYPE = auth_type or kwargs.get("AUTH_TYPE", AuthTypes.client_secret)
-        self.CLIENT_ID = client_id or kwargs.get("CLIENT_ID", None)
-        self.AUTHORITY = authority or kwargs.get("AUTHORITY", None)
-        self.TENANT_ID = tenant_id or kwargs.get("TENANT_ID", None)
-        self.CLIENT_SECRET = client_secret or kwargs.get("CLIENT_SECRET", None)
-        self.CERT_PEM_FILE = cert_pem_file or kwargs.get("CERT_PEM_FILE", None)
-        self.CERT_KEY_FILE = cert_key_file or kwargs.get("CERT_KEY_FILE", None)
-        self.CONNECTION_NAME = connection_name or kwargs.get("CONNECTION_NAME", None)
-        self.SCOPES = scopes or kwargs.get("SCOPES", None)
