+++ libraries/Core/microsoft-agents-authorization/microsoft/agents/authorization/auth_types.py
