+++ libraries/Hosting/microsoft-agents-hosting-aiohttp/microsoft/agents/hosting/aiohttp/cloud_adapter.py
@@ -23,6 +23,7 @@ from microsoft.agents.builder import (
     ChannelServiceAdapter,
     ChannelServiceClientFactoryBase,
     MessageFactory,
+    RestChannelServiceClientFactory,
     TurnContext,
 )
 
@@ -33,8 +34,7 @@ class CloudAdapter(ChannelServiceAdapter, AgentHttpAdapter):
     def __init__(
         self,
         *,
-        service_auth_configuration: AgentAuthConfiguration = None,
-        authorization: Authorization = None,
+        connection_manager: Connections,
         channel_service_client_factory: ChannelServiceClientFactoryBase = None,
     ):
         """
@@ -42,9 +42,6 @@ class CloudAdapter(ChannelServiceAdapter, AgentHttpAdapter):
 
         :param channel_service_client_factory: The factory to use to create the channel service client.
         """
-        
-        super().__init__(channel_service_client_factory)
-
         async def on_turn_error(context: TurnContext, error: Exception):
             error_message = f"Exception caught : {error}"
             print(format_exc())
@@ -60,7 +57,11 @@ class CloudAdapter(ChannelServiceAdapter, AgentHttpAdapter):
             )
 
         self.on_turn_error = on_turn_error
-        self._channel_service_client_factory = channel_service_client_factory
+
+        #TODO: Implement creating a connection manager out of environment variables
+        self._channel_service_client_factory = channel_service_client_factory or RestChannelServiceClientFactory(connection_manager)
+
+        super().__init__(channel_service_client_factory)
 
     async def process(self, request: Request, agent: Agent) -> Optional[Response]:
         if not request:
